+++ glean-core/ffi/src/uuid.rs
@@ -13,6 +13,7 @@ use crate::{
 
 define_metric!(UuidMetric => UUID_METRICS {
     new           -> glean_new_uuid_metric(),
+    test_get_num_recorded_errors -> glean_uuid_test_get_num_recorded_errors,
     destroy       -> glean_destroy_uuid_metric,
 });
 
@@ -21,14 +22,7 @@ pub extern "C" fn glean_uuid_set(metric_id: u64, value: FfiStr) {
     with_glean_value(|glean| {
         UUID_METRICS.call_with_log(metric_id, |metric| {
             let value = value.to_string_fallible()?;
-            if let Ok(uuid) = uuid::Uuid::parse_str(&value) {
-                metric.set(glean, uuid);
-            } else {
-                log::error!(
-                    "Unexpected `uuid` value coming from platform code '{}'",
-                    value
-                );
-            }
+            metric.set_from_str(glean, &value);
             Ok(())
         })
     })