+++ glean-core/src/metrics/uuid.rs
@@ -4,6 +4,7 @@
 
 use uuid::Uuid;
 
+use crate::error_recording::{record_error, ErrorType};
 use crate::metrics::Metric;
 use crate::metrics::MetricType;
 use crate::storage::StorageManager;
@@ -52,6 +53,28 @@ impl UuidMetric {
         let s = value.to_string();
         let value = Metric::Uuid(s);
         glean.storage().record(glean, &self.meta, &value)
+    }
+
+    /// Sets to the specified value, from a string.
+    ///
+    /// This should only be used from FFI. When calling directly from Rust, it
+    /// is better to use `set`.
+    ///
+    /// # Arguments
+    ///
+    /// * `glean` - The Glean instance this metric belongs to.
+    /// * `value` - The UUID to set the metric to.
+    pub fn set_from_str(&self, glean: &Glean, value: &str) {
+        if !self.should_record(glean) {
+            return;
+        }
+
+        if let Ok(uuid) = uuid::Uuid::parse_str(&value) {
+            self.set(glean, uuid);
+        } else {
+            let msg = format!("Unexpected UUID value '{}'", value);
+            record_error(glean, &self.meta, ErrorType::InvalidValue, msg, None);
+        }
     }
 
     /// Generates a new random UUID and set the metric to it.