+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Rust
   * Introduce the UUID metric type in the RLB.
   * Introduce the Labeled metric type in the RLB [#1327](https://github.com/mozilla/glean/pull/1327).
+* Python
+  * BUGFIX: Setting a UUID metric to a value that is not in the expected UUID format will now record an error with the Glean error reporting system.
 
 # v33.4.0 (2020-11-17)